created two components for email and sms - pf
