+++ packages/pn-personafisica-webapp/public/locales/en/recapiti.json
@@ -9,8 +9,8 @@
   "subtitle-text-4": ".",
   "general-contacts-title": "General contact details",
   "special-contacts-title": "Other contact details",
-  "alert-dialog-EMAIL": "If you do not have a PEC and access the notification on SEND within 5 days (120 hours) of sending the email, you will not receive the acknowledgement of receipt by registered mail.",
-  "alert-dialog-SMS": "If you do not have a PEC and access the SEND notification within 5 days (120 hours) of sending the SMS, you will not receive the acknowledgement of receipt by registered mail.",
+  "alert-dialog-email": "If you do not have a PEC and access the notification on SEND within 5 days (120 hours) of sending the email, you will not receive the acknowledgement of receipt by registered mail.",
+  "alert-dialog-sms": "If you do not have a PEC and access the SEND notification within 5 days (120 hours) of sending the SMS, you will not receive the acknowledgement of receipt by registered mail.",
   "legal-contacts": {
     "title": "",
     "subtitle": "Legal address",
