+++ packages/pn-personafisica-webapp/public/locales/fr/recapiti.json
@@ -9,8 +9,8 @@
   "subtitle-text-4": ".",
   "general-contacts-title": "Coordonnées générales",
   "special-contacts-title": "Autres coordonnées",
-  "alert-dialog-EMAIL": "Si vous n’avez pas de PEC et que vous accédez à la notification sur SEND dans les 5 jours (120 heures) suivant l’envoi de l’e-mail, vous ne recevrez pas d’avis de réception par lettre recommandée.",
-  "alert-dialog-SMS": "Si vous n’avez pas de PEC et que vous accédez à la notification sur SEND dans les 5 jours (120 heures) suivant l’envoi du SMS, vous ne recevrez pas d’avis de réception par lettre recommandée.",
+  "alert-dialog-email": "Si vous n’avez pas de PEC et que vous accédez à la notification sur SEND dans les 5 jours (120 heures) suivant l’envoi de l’e-mail, vous ne recevrez pas d’avis de réception par lettre recommandée.",
+  "alert-dialog-sms": "Si vous n’avez pas de PEC et que vous accédez à la notification sur SEND dans les 5 jours (120 heures) suivant l’envoi du SMS, vous ne recevrez pas d’avis de réception par lettre recommandée.",
   "legal-contacts": {
     "title": "",
     "subtitle": "Coordonnées légales",
