+++ packages/pn-personafisica-webapp/public/locales/it/recapiti.json
@@ -9,8 +9,8 @@
   "subtitle-text-4": ".",
   "general-contacts-title": "Recapiti generali",
   "special-contacts-title": "Altri recapiti",
-  "alert-dialog-EMAIL": "Se non hai una PEC e accedi alla notifica su SEND entro 5 giorni (120 ore) dall’invio dell'e-mail, non riceverai l'avviso di avvenuta ricezione tramite raccomandata.",
-  "alert-dialog-SMS": "Se non hai una PEC e accedi alla notifica su SEND entro 5 giorni (120 ore) dall’invio dell'SMS, non riceverai l'avviso di avvenuta ricezione tramite raccomandata.",
+  "alert-dialog-email": "Se non hai una PEC e accedi alla notifica su SEND entro 5 giorni (120 ore) dall’invio dell'e-mail, non riceverai l'avviso di avvenuta ricezione tramite raccomandata.",
+  "alert-dialog-sms": "Se non hai una PEC e accedi alla notifica su SEND entro 5 giorni (120 ore) dall’invio dell'SMS, non riceverai l'avviso di avvenuta ricezione tramite raccomandata.",
   "legal-contacts": {
     "title": "",
     "subtitle": "Recapito legale",
