+++ packages/pn-personafisica-webapp/src/components/Contacts/SmsContactItem.tsx
@@ -3,12 +3,12 @@ import { ChangeEvent, useCallback, useEffect, useRef, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
 import * as yup from 'yup';
 
-import WatchLaterIcon from '@mui/icons-material/WatchLater';
-import { Box, Button, Stack, TextField, Typography } from '@mui/material';
+import { Box, Button, InputAdornment, Stack, TextField, Typography } from '@mui/material';
 import {
   AppResponse,
   AppResponsePublisher,
   CodeModal,
+  DisclaimerModal,
   ErrorMessage,
   appStateActions,
 } from '@pagopa-pn/pn-commons';
@@ -21,28 +21,28 @@ import { SaveDigitalAddressParams } from '../../redux/contact/types';
 import { useAppDispatch, useAppSelector } from '../../redux/hooks';
 import { RootState } from '../../redux/store';
 import PFEventStrategyFactory from '../../utility/MixpanelUtils/PFEventStrategyFactory';
-import { contactAlreadyExists, pecValidationSchema } from '../../utility/contacts.utility';
-import CancelVerificationModal from './CancelVerificationModal';
+import {
+  contactAlreadyExists,
+  internationalPhonePrefix,
+  phoneValidationSchema,
+} from '../../utility/contacts.utility';
 import DeleteDialog from './DeleteDialog';
 import DigitalContactElem from './DigitalContactElem';
 import ExistingContactDialog from './ExistingContactDialog';
-import PecVerificationDialog from './PecVerificationDialog';
 
-type Props = {
+interface Props {
   value: string;
-  verifyingAddress: boolean;
   blockDelete?: boolean;
-};
+}
 
 enum ModalType {
   EXISTING = 'existing',
-  VALIDATION = 'validation',
-  CANCEL_VALIDATION = 'cancel_validation',
-  DELETE = 'delete',
+  DISCLAIMER = 'disclaimer',
   CODE = 'code',
+  DELETE = 'delete',
 }
 
-const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
+const SmsContactItem = ({ value, blockDelete }: Props) => {
   const { t } = useTranslation(['common', 'recapiti']);
   const digitalAddresses =
     useAppSelector((state: RootState) => state.contactsState.digitalAddresses) ?? [];
@@ -55,12 +55,12 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
   const codeModalRef =
     useRef<{ updateError: (error: ErrorMessage, codeNotValid: boolean) => void }>(null);
 
-  const validationSchema = yup.object({
-    pec: pecValidationSchema(t),
+  const validationSchema = yup.object().shape({
+    sms: phoneValidationSchema(t, !!value),
   });
 
   const initialValues = {
-    pec: value,
+    sms: value ?? '',
   };
 
   const formik = useFormik({
@@ -68,17 +68,19 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
     validationSchema,
     validateOnMount: true,
     enableReinitialize: true,
-    /** onSubmit validate */
     onSubmit: () => {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_PEC_START, 'default');
+      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SMS_START, 'default');
       // first check if contact already exists
-      if (contactAlreadyExists(digitalAddresses, formik.values.pec, 'default', ChannelType.PEC)) {
+      if (contactAlreadyExists(digitalAddresses, getContactValue(), 'default', ChannelType.SMS)) {
         setModalOpen(ModalType.EXISTING);
         return;
       }
-      handleCodeVerification();
+      setModalOpen(ModalType.DISCLAIMER);
     },
   });
+
+  const getContactValue = () =>
+    value ? formik.values.sms : internationalPhonePrefix + formik.values.sms;
 
   const handleChangeTouched = async (e: ChangeEvent) => {
     formik.handleChange(e);
@@ -87,14 +89,14 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
 
   const handleCodeVerification = (verificationCode?: string) => {
     if (verificationCode) {
-      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_PEC_UX_CONVERSION, 'default');
+      PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SMS_UX_CONVERSION, 'default');
     }
 
     const digitalAddressParams: SaveDigitalAddressParams = {
-      addressType: AddressType.LEGAL,
+      addressType: AddressType.COURTESY,
       senderId: 'default',
-      channelType: ChannelType.PEC,
-      value: formik.values.pec,
+      channelType: ChannelType.SMS,
+      value: getContactValue(),
       code: verificationCode,
     };
 
@@ -104,30 +106,29 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
         // contact to verify
         // open code modal
         if (!res) {
+          // aprire la code modal
           setModalOpen(ModalType.CODE);
           return;
         }
 
-        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_PEC_UX_SUCCESS, 'default');
+        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SMS_UX_SUCCESS, 'default');
 
         // contact has already been verified
-        if (res.pecValid) {
-          // show success message
-          dispatch(
-            appStateActions.addSuccess({
-              title: '',
-              message: t(`legal-contacts.pec-added-successfully`, { ns: 'recapiti' }),
-            })
-          );
-          setModalOpen(null);
-          if (value) {
-            digitalElemRef.current.toggleEdit();
-          }
-          return;
+        // show success message
+        dispatch(
+          appStateActions.addSuccess({
+            title: '',
+            message: t(`courtesy-contacts.sms-added-successfully`, {
+              ns: 'recapiti',
+            }),
+          })
+        );
+        // chiudere la code modal
+        setModalOpen(null);
+        // nel caso siamo in modifica (property value defined), bisogna passare alla modalità noEdit del componente DigitalContactElem
+        if (value) {
+          digitalElemRef.current.toggleEdit();
         }
-        // contact must be validated
-        // open validation modal
-        setModalOpen(ModalType.VALIDATION);
       })
       .catch(() => {});
   };
@@ -137,22 +138,22 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
     if (value) {
       digitalElemRef.current.toggleEdit();
     }
-    await formik.setFieldTouched('pec', false, false);
-    await formik.setFieldValue('pec', initialValues.pec, true);
+    await formik.setFieldTouched('sms', false, false);
+    await formik.setFieldValue('sms', initialValues.sms, true);
   };
 
   const deleteConfirmHandler = () => {
     setModalOpen(null);
     dispatch(
       deleteAddress({
-        addressType: AddressType.LEGAL,
+        addressType: AddressType.COURTESY,
         senderId: 'default',
-        channelType: ChannelType.PEC,
+        channelType: ChannelType.SMS,
       })
     )
       .unwrap()
       .then(() => {
-        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_REMOVE_PEC_SUCCESS, 'default');
+        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_REMOVE_SMS_SUCCESS, 'default');
       })
       .catch(() => {});
   };
@@ -172,7 +173,7 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
           },
           true
         );
-        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_PEC_CODE_ERROR);
+        PFEventStrategyFactory.triggerEvent(PFEventsType.SEND_ADD_SMS_CODE_ERROR);
       }
       return false;
     },
@@ -188,107 +189,103 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
   }, [handleAddressUpdateError]);
 
   /*
-   * if *some* value has been attached to the contact type,
+   * if *some* value (phone number, email address) has been attached to the contact type,
    * then we show the value giving the user the possibility of changing it
    * (the DigitalContactElem component includes the "update" button)
-   */
-  /*
-   * if *no* value has been attached to the contact type,
+   * if *no* value (phone number, email address) has been attached to the contact type,
    * then we show the input field allowing the user to enter it along with the button
    * to perform the addition.
    */
   return (
     <>
-      <form onSubmit={formik.handleSubmit} data-testid="pecContact">
-        {value && (
-          <>
-            <Typography mb={1} sx={{ fontWeight: 'bold' }} id="associatedPEC" mt={3}>
-              {t('legal-contacts.pec-added', { ns: 'recapiti' })}
-            </Typography>
-            <DigitalContactElem
-              senderId="default"
-              contactType={ChannelType.PEC}
-              ref={digitalElemRef}
-              inputProps={{
-                id: 'pec',
-                name: 'pec',
-                label: 'PEC',
-                value: formik.values.pec,
-                onChange: (e) => void handleChangeTouched(e),
-                error: formik.touched.pec && Boolean(formik.errors.pec),
-                helperText: formik.touched.pec && formik.errors.pec,
-              }}
-              saveDisabled={!formik.isValid}
-              onDelete={() => setModalOpen(ModalType.DELETE)}
-              onEditCancel={() => formik.resetForm({ values: initialValues })}
-              editManagedFromOutside
-            />
-          </>
-        )}
-        {verifyingAddress && (
-          <>
-            <Typography mb={1} sx={{ fontWeight: 'bold' }} mt={3}>
-              {t('legal-contacts.pec-validating', { ns: 'recapiti' })}
-            </Typography>
-            <Stack direction="row" spacing={1}>
-              <WatchLaterIcon fontSize="small" />
-              <Typography id="validationPecProgress" fontWeight="bold" variant="body2">
-                {t('legal-contacts.validation-in-progress', { ns: 'recapiti' })}
-              </Typography>
-              <ButtonNaked
-                color="primary"
-                onClick={() => setModalOpen(ModalType.CANCEL_VALIDATION)}
-                data-testid="cancelValidation"
-              >
-                {t('legal-contacts.cancel-pec-validation', { ns: 'recapiti' })}
-              </ButtonNaked>
-            </Stack>
-          </>
-        )}
-        {!value && !verifyingAddress && (
-          <Stack spacing={2} direction={{ sm: 'row', xs: 'column' }} mt={3}>
+      <form onSubmit={formik.handleSubmit} data-testid="courtesyContacts-sms">
+        <Typography id="sms-label" variant="body2" mb={1} sx={{ fontWeight: 'bold' }}>
+          {t(`courtesy-contacts.sms-added`, { ns: 'recapiti' })}
+        </Typography>
+        {value ? (
+          <DigitalContactElem
+            senderId="default"
+            contactType={ChannelType.SMS}
+            ref={digitalElemRef}
+            inputProps={{
+              id: 'sms',
+              name: 'sms',
+              label: t(`courtesy-contacts.link-sms-placeholder`, {
+                ns: 'recapiti',
+              }),
+              value: formik.values.sms,
+              onChange: (e) => void handleChangeTouched(e),
+              error: formik.touched.sms && Boolean(formik.errors.sms),
+              helperText: formik.touched.sms && formik.errors.sms,
+            }}
+            saveDisabled={!formik.isValid}
+            onDelete={() => setModalOpen(ModalType.DELETE)}
+            onEditCancel={() => formik.resetForm({ values: initialValues })}
+            editManagedFromOutside
+          />
+        ) : (
+          <Stack direction={{ xs: 'column', sm: 'row' }} spacing={2}>
             <TextField
-              id="pec"
-              placeholder={t('legal-contacts.link-pec-placeholder', { ns: 'recapiti' })}
-              fullWidth
-              name="pec"
-              value={formik.values.pec}
+              id="sms"
+              name="sms"
+              value={formik.values.sms}
               onChange={handleChangeTouched}
-              error={formik.touched.pec && Boolean(formik.errors.pec)}
-              helperText={formik.touched.pec && formik.errors.pec}
+              error={formik.touched.sms && Boolean(formik.errors.sms)}
+              helperText={formik.touched.sms && formik.errors.sms}
               inputProps={{ sx: { height: '14px' } }}
+              placeholder={t(`courtesy-contacts.link-sms-placeholder`, {
+                ns: 'recapiti',
+              })}
+              fullWidth
+              InputProps={{
+                startAdornment: (
+                  <InputAdornment position="start">{internationalPhonePrefix}</InputAdornment>
+                ),
+              }}
               sx={{ flexBasis: { xs: 'unset', lg: '66.66%' } }}
             />
+
             <Button
-              id="add-contact"
+              id="courtesy-sms-button"
               variant="outlined"
               disabled={!formik.isValid}
               fullWidth
               type="submit"
-              data-testid="addContact"
+              data-testid="courtesy-sms-button"
               sx={{ flexBasis: { xs: 'unset', lg: '33.33%' } }}
             >
-              {t('button.conferma')}
+              {t(`courtesy-contacts.sms-add`, { ns: 'recapiti' })}
             </Button>
           </Stack>
         )}
       </form>
       <ExistingContactDialog
         open={modalOpen === ModalType.EXISTING}
-        value={formik.values.pec}
+        value={formik.values.sms}
         handleDiscard={() => setModalOpen(null)}
         handleConfirm={() => handleCodeVerification()}
       />
+      <DisclaimerModal
+        open={modalOpen === ModalType.DISCLAIMER}
+        onConfirm={() => {
+          setModalOpen(null);
+          handleCodeVerification();
+        }}
+        onCancel={() => setModalOpen(null)}
+        confirmLabel={t('button.conferma')}
+        checkboxLabel={t('button.capito')}
+        content={t(`alert-dialog-sms`, { ns: 'recapiti' })}
+      />
       <CodeModal
-        title={t(`legal-contacts.pec-verify`, { ns: 'recapiti' }) + ` ${formik.values.pec}`}
-        subtitle={<Trans i18nKey={`legal-contacts.pec-verify-descr`} ns="recapiti" />}
+        title={t(`courtesy-contacts.sms-verify`, { ns: 'recapiti' }) + ` ${formik.values.sms}`}
+        subtitle={<Trans i18nKey={`courtesy-contacts.sms-verify-descr`} ns="recapiti" />}
         open={modalOpen === ModalType.CODE}
         initialValues={new Array(5).fill('')}
-        codeSectionTitle={t(`legal-contacts.insert-code`, { ns: 'recapiti' })}
+        codeSectionTitle={t(`courtesy-contacts.insert-code`, { ns: 'recapiti' })}
         codeSectionAdditional={
           <>
             <Typography variant="body2" display="inline">
-              {t(`legal-contacts.pec-new-code`, { ns: 'recapiti' })}
+              {t(`courtesy-contacts.sms-new-code`, { ns: 'recapiti' })}
               &nbsp;
             </Typography>
             <ButtonNaked
@@ -302,7 +299,7 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
                 variant="body2"
                 sx={{ textDecoration: 'underline' }}
               >
-                {t(`legal-contacts.new-code-link`, { ns: 'recapiti' })}.
+                {t(`courtesy-contacts.new-code-link`, { ns: 'recapiti' })}.
               </Typography>
             </ButtonNaked>
           </>
@@ -313,21 +310,13 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
         confirmCallback={(values: Array<string>) => handleCodeVerification(values.join(''))}
         ref={codeModalRef}
       />
-      <PecVerificationDialog
-        open={modalOpen === ModalType.VALIDATION}
-        handleConfirm={() => setModalOpen(null)}
-      />
-      <CancelVerificationModal
-        open={modalOpen === ModalType.CANCEL_VALIDATION}
-        handleClose={() => setModalOpen(null)}
-      />
       <DeleteDialog
         showModal={modalOpen === ModalType.DELETE}
-        removeModalTitle={t(`legal-contacts.${blockDelete ? 'block-' : ''}remove-pec-title`, {
+        removeModalTitle={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-sms-title`, {
           ns: 'recapiti',
         })}
-        removeModalBody={t(`legal-contacts.${blockDelete ? 'block-' : ''}remove-pec-message`, {
-          value: formik.values.pec,
+        removeModalBody={t(`courtesy-contacts.${blockDelete ? 'block-' : ''}remove-sms-message`, {
+          value: formik.values.sms,
           ns: 'recapiti',
         })}
         handleModalClose={() => setModalOpen(null)}
@@ -338,4 +327,4 @@ const PecContactItem = ({ value, verifyingAddress, blockDelete }: Props) => {
   );
 };
 
-export default PecContactItem;
+export default SmsContactItem;
